tests(s3): split into dedicated test files
