+++ packages/s3/test/utils.js
@@ -1,0 +1,13 @@
+// @ts-check
+import { rimraf } from 'rimraf'
+
+export const testResultsDirectory = './test/results'
+
+/**
+ * Remove the test results directory.
+ *
+ * @returns {Promise<void>}
+ */
+export const removeResultsDirectory = async () => {
+  await rimraf(testResultsDirectory)
+}
